Add files via upload
Here is just a start to the code so we can all start working on it. I just noticed that in the UML diagram we have we have its a steak timer and I make the time constraints for a boiled egg. We can totally change this or even do both were the user has an option of the food they want to cook. Whatever everyone wants to do and how much time we have. :)